Fix font-sizing
Somehow the body set the font-size to 1.125, which left the body font at
a non-integer value.  Removed this, which then made the average
characters/line too large, so then reduced the max-width-container's
width.  Also added another reponsive level, since the largest font-size
was decreased when the 1.125 body font-size was changed.